Cleanup encryption again by files
